Revert "reverting blocking shield and print event listeners, and putting them into classes."
This reverts commit 8b08d032e5aacd8406e6990151ec777837ce3b03.
